Refactored file structure and code relatively
